Refactored default account image to render as part of the UserAvatar instead of part of the PageHeader. #8
